Feat/#110/멀티파트 폼 데이터 에러 처리 (#111)
* [feat] QuizService 에서 이미지가 안 들어왔을 때에 대한 처리

* [feat] drawService 에서 이미지가 안 들어왔을 때에 대한 처리

* [feat] drawRewardInfoService, drawProbabilityService 삭제 후 저장 로직 추가

* [feat] adminController 예외처리 로직 추가

* [feat] controller가 예외를 던지지 않게 수정

---------

+++ src/main/java/softeer/team_pineapple_be/domain/quiz/service/QuizService.java
@@ -167,6 +167,13 @@ public class QuizService {
       return;
     }
     QuizInfo quizInfo = quizInfoByDate.get();
+    QuizContent quizContent = quizContentRepository.findByQuizDate(day)
+            .orElseThrow(
+                    () -> new RestApiException(QuizErrorCode.NO_QUIZ_CONTENT));
+    if(quizInfoModifyRequest.getQuizImage() == null){
+      quizInfoRepository.save(new QuizInfo(quizInfo.getId(), quizContent, quizInfoModifyRequest.getAnswerNum(), quizInfo.getQuizImage()));
+      return;
+    }
     s3DeleteService.deleteFolder(fileName);
     imageUrl = uploadImageToS3(quizInfoModifyRequest, fileName);
     quizInfo.update(quizInfoModifyRequest.getAnswerNum(), imageUrl);
@@ -275,6 +282,8 @@ public class QuizService {
     try {
       imageUrl = s3UploadService.saveFile(quizInfoModifyRequest.getQuizImage(), fileName);
     } catch (IOException e) {
+      throw new RestApiException(QuizErrorCode.NO_QUIZ_IMAGE);
+    } catch (RuntimeException e) {
       throw new RestApiException(QuizErrorCode.NO_QUIZ_IMAGE);
     }
     return imageUrl;
